+++ demisto_sdk/common/hook_validations/structure.py
@@ -10,10 +10,8 @@ from typing import Optional
 import yaml
 from pykwalify.core import Core
 
-from demisto_sdk.common.constants import YML_INTEGRATION_REGEXES, YML_ALL_PLAYBOOKS_REGEX, YML_SCRIPT_REGEXES, \
-    JSON_ALL_WIDGETS_REGEXES, JSON_ALL_DASHBOARDS_REGEXES, JSON_ALL_CONNECTIONS_REGEXES, JSON_ALL_CLASSIFIER_REGEXES, \
-    JSON_ALL_LAYOUT_REGEXES, JSON_ALL_INCIDENT_FIELD_REGEXES, JSON_ALL_REPORTS_REGEXES, MISC_REPUTATIONS_REGEX, \
-    MISC_REGEX, Errors, ACCEPTED_FILE_EXTENSIONS
+from demisto_sdk.common.constants import Errors, ACCEPTED_FILE_EXTENSIONS, FILE_TYPES_PATHS_TO_VALIDATE, \
+    SCHEMA_TO_REGEX, REPUTATION_REGEX
 from demisto_sdk.common.tools import get_remote_file, get_matching_regex, print_error
 from demisto_sdk.common.configuration import Configuration
 
@@ -36,26 +34,9 @@ class StructureValidator:
         '.yml': yaml.safe_load,
         '.json': json.load,
     }
-    SCHEMA_TO_REGEX = {
-        'integration': YML_INTEGRATION_REGEXES,
-        'playbook': YML_ALL_PLAYBOOKS_REGEX,
-        'script': YML_SCRIPT_REGEXES,
-        'widget': JSON_ALL_WIDGETS_REGEXES,
-        'dashboard': JSON_ALL_DASHBOARDS_REGEXES,
-        'canvas-context-connections': JSON_ALL_CONNECTIONS_REGEXES,
-        'classifier': JSON_ALL_CLASSIFIER_REGEXES,
-        'layout': JSON_ALL_LAYOUT_REGEXES,
-        'incidentfield': JSON_ALL_INCIDENT_FIELD_REGEXES,
-    }
-
-    PATHS_TO_VALIDATE = {
-        'reports': JSON_ALL_REPORTS_REGEXES,
-        'reputation': [MISC_REPUTATIONS_REGEX],
-        'reputations': [MISC_REGEX]
-    }
 
     def __init__(self, file_path, old_file_path=None, predefined_scheme=None, configuration=Configuration()):
-        # type: (str, Optional[str], Configuration) -> None
+        # type: (str, Optional[str], Optional[str], Configuration) -> None
         self.is_valid = None  # type: Optional[bool]
         self.file_path = file_path
         self.scheme_name = predefined_scheme or self.scheme_of_file_by_path()
@@ -80,6 +61,7 @@ class StructureValidator:
         if self.old_file:  # In case the file is modified
             answers.append(not self.is_id_modified())
             answers.append(self.is_valid_fromversion_on_modified())
+
         return all(answers)
 
     def scheme_of_file_by_path(self):
@@ -89,11 +71,15 @@ class StructureValidator:
         Returns:
             (str): Type of file by scheme name
         """
-        for scheme_name, regex_list in self.SCHEMA_TO_REGEX.items():
+
+        for scheme_name, regex_list in SCHEMA_TO_REGEX.items():
             if get_matching_regex(self.file_path, regex_list):
                 return scheme_name
 
-        pretty_formated_string_of_regexes = json.dumps(self.SCHEMA_TO_REGEX, indent=4, sort_keys=True)
+        if get_matching_regex(self.file_path, [REPUTATION_REGEX]):
+            return 'reputation'
+
+        pretty_formated_string_of_regexes = json.dumps(SCHEMA_TO_REGEX, indent=4, sort_keys=True)
 
         print_error(f"The file {self.file_path} does not match any scheme we have please, refer to the following list"
                     f" for the various file name options we have in our repo {pretty_formated_string_of_regexes}")
@@ -106,7 +92,7 @@ class StructureValidator:
         Returns:
             bool. Whether the scheme is valid on self.file_path.
         """
-        if self.scheme_name is None:
+        if self.scheme_name in [None, 'reputation', 'image']:
             return True
         try:
             path = os.path.normpath(
@@ -172,7 +158,12 @@ class StructureValidator:
 
         old_version_id = self.get_file_id_from_loaded_file_data(self.old_file)
         new_file_id = self.get_file_id_from_loaded_file_data(self.current_file)
-        return not (new_file_id == old_version_id)
+        if not (new_file_id == old_version_id):
+            print_error(f"The file id for {self.file_path} has changed from {old_version_id} to {new_file_id}")
+            return True
+
+        # False - the id has not changed.
+        return False
 
     def is_valid_fromversion_on_modified(self):
         # type: () -> bool
@@ -208,6 +199,10 @@ class StructureValidator:
                     loaded_file_data = load_function(file_obj)  # type: ignore
                     return loaded_file_data
 
+            # Ignore loading image
+            elif file_extension == '.png':
+                return {}
+
         print_error(Errors.wrong_file_extension(file_extension, self.FILE_SUFFIX_TO_LOAD_FUNCTION.keys()))
         return {}
 
@@ -222,7 +217,7 @@ class StructureValidator:
         if self.scheme_name:
             return self.scheme_name
 
-        for file_type, regexes in self.PATHS_TO_VALIDATE.items():
+        for file_type, regexes in FILE_TYPES_PATHS_TO_VALIDATE.items():
             for regex in regexes:
                 if re.search(regex, self.file_path, re.IGNORECASE):
                     return file_type