+++ CHANGELOG.md
@@ -4,6 +4,27 @@
 
 [1]: https://pypi.org/project/demisto-sdk/#history
 
+### 0.3.4
+* Fixed an issue where "_test" file for scripts/integrations created using **init** would import the "HelloWorld" templates.
+* Add validation for docker image inside integrations and scripts.
+* Added **spell-check** command.
+
+### 0.3.3
+* Added backwards compatibility break error message.
+* Added schema for incident types.
+* Added **additionalinfo** field to as an available field for integration configuration.
+* Added pack parameter for **init**.
+* Fixed an issue where error would appear if name parameter is not set in **init**.
+
+### 0.3.2
+* Fixed the handling of classifier files in **validate**.
+
+
+### 0.3.1
+* Fixed the handling of newly created reputation files in **validate**.
+* Added an option to perform **validate** on a specific file.
+
+
 ### 0.3.0
 
 * Added support for multi-package **lint** both with parallel and without.
@@ -13,7 +34,18 @@
     * Playbooks
     * Integrations
 * Improved user outputs for **secrets** command.
-* Added **spell-check** command.
+* Fixed an issue where **lint** would run pytest and pylint only on a single docker per integration.
+* Added auto-complete functionality to demisto-sdk.
+* Added git parameter in **lint** to run only on changed packages.
+* Added the **run-playbook** command
+* Added **run** command which runs a command in the Demisto playground.
+* Added **upload** command which uploads an integration or a script to a Demisto instance.
+* Fixed and issue where **validate** checked if release notes exist for new integrations and scripts.
+* Added **generate-test-playbook** command which generates a basic test playbook for an integration or a script.
+* **validate** now supports indicator fields.
+* Fixed an issue with layouts scheme validation.
+* Adding **init** command.
+* Added **json-to-outputs** command which generates the yaml section for outputs from an API raw response.
 
 ### 0.2.6